requirements.txt
streamlit==1.45.1
sdmx1==1.1.0
